requirements.txt
deeplake==3.4.1
google-search-results==2.4.2
langchain==0.0.165
openai==0.27.6
PyGithub==1.58.2
python-dotenv==1.0.0
tiktoken==0.4.0